fix bug

+++ depoly.sh
@@ -9,27 +9,30 @@ sed -i "s/YOUR_PROJECT_NAME/$projectName/g" docker-compose.yml
 sed -i "s/YOUR_PROJECT_DOMAIN.com/$projectDomain/g" docker-compose.yml
 
 DEFAULT="y"
-read -p "Enable redis? [Y/n]:" eRedis
+read -p "Enable redis? [Y/n]: " eRedis
 eRedis="${eRedis:-${DEFAULT}}"
 
-if [ "${eRedis}" = "y" ] && [ "${eRedis}" = "Y" ]; then
+if [ "${eRedis}" = "y" ] || [ "${eRedis}" = "Y" ]; then
 	sed -i "s/#redis//g" docker-compose.yml
 fi
 
-read -p "Enable meilisearch? [Y/n]:" eMeilisearch
+read -p "Enable meilisearch? [Y/n]: " eMeilisearch
 eMeilisearch="${eRedis:-${DEFAULT}}"
 
 
-if [ "${eMeilisearch}" = "y" ] && [ "${eMeilisearch}" = "Y" ] ; then
+if [ "${eMeilisearch}" = "y" ] || [ "${eMeilisearch}" = "Y" ] ; then
 	sed -i "s/#meilisearch//g" docker-compose.yml
 fi
 
-cp ./src/.env.example ./src/.env
-sed -i "s/DB_HOST=.*/DB_HOST=mysql/g" ./src/.env
-sed -i "s/DB_PASSWORD=.*/DB_PASSWORD=password/g" ./src/.env
-sed -i "s/APP_NAME=.*/APP_NAME=$projectName/g" ./src/.env
-sed -i "s/APP_ENV=.*/APP_ENV=production/g" ./src/.env
-sed -i "s/APP_URL=.*/APP_URL=https:\/\/$projectDomain/g" ./src/.env
+FILE=./src/.env.example
+if test -f "$FILE"; then
+    cp ./src/.env.example ./src/.env
+	sed -i "s/DB_HOST=.*/DB_HOST=mysql/g" ./src/.env
+	sed -i "s/DB_PASSWORD=.*/DB_PASSWORD=password/g" ./src/.env
+	sed -i "s/APP_NAME=.*/APP_NAME=$projectName/g" ./src/.env
+	sed -i "s/APP_ENV=.*/APP_ENV=production/g" ./src/.env
+	sed -i "s/APP_URL=.*/APP_URL=https:\/\/$projectDomain/g" ./src/.env
+fi
 
 cp ./backup/backup_db.sh.example backup_db.sh
 sed -i "s/YOUR_PROJECT_NAME/$projectName/g" backup_db.sh
